requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
asyncpg==0.30.0
gunicorn==23.0.0
Werkzeug==3.1.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
Flask-Mail==0.9.1
python-dotenv==1.0.0
Flask-Migrate==4.0.5
APScheduler==3.10.4
Flask-WTF==1.2.1
WTForms==3.1.1
